Switch to Maven and use Java SDK for Watson services.
