+++ .gitignore
@@ -1,6 +1,4 @@
 bin/
-WebContent/WEB-INF/classes
-build/
-/output/webApp.war
 .DS_Store
-output/
+/target
+*~
