--- .settings/org.eclipse.wst.ws.service.policy.prefs
@@ -1,3 +1,0 @@
-com.ibm.ast.ws.jaxws.annotations.v7.jaxws.default.value.key=com.ibm.ast.ws.jaxws.annotations.v7.default
-eclipse.preferences.version=1
-org.eclipse.wst.ws.service.policy.projectEnabled=false
