--- lib/README.txt
@@ -1,2 +1,0 @@
-# lib folder
-  The lib folder contains the libraries to build the war file, see build.xml
