+++ src/main/webapp/js/browser-detect.js
