+++ src/main/webapp/js/constants.js
